eloworld/requirements.txt
django==1.11
gunicorn==19.7.1
django-field-history==0.6.0